requirements.txt
haystack-ai
jq
"sentence-transformers>=5.0.0"
Flask
